requirements.txt
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
beautifulsoup4
certifi
charset-normalizer
click
coloredlogs
dataclasses-json
Deprecated
dirtyjson
diskcache
distro
exceptiongroup
fastembed
filelock
flatbuffers
frozenlist
fsspec
greenlet
grpcio
grpcio-tools
h11
h2
hpack
httpcore
httpx
huggingface-hub
humanfriendly
hyperframe
idna
Jinja2
joblib
jsonpatch
jsonpointer
langchain-core
langchain-openai
langsmith
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-fastembed
llama-index-embeddings-openai
llama-index-graph-stores-neo4j
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-llama-cpp
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-vector-stores-neo4jvector
llama-index-vector-stores-qdrant
llama-parse
llama_cpp_python
llamaindex-py-client
loguru
MarkupSafe
marshmallow
mpmath
multidict
mypy-extensions
neo4j
nest-asyncio
networkx
nltk
numpy
onnx
onnxruntime
openai
orjson
packaging
pandas
pillow
portalocker
protobuf
pydantic
pydantic_core
pypdf
python-dateutil
python-dotenv
pytz
PyYAML
qdrant-client
regex
requests
scikit-learn
scipy
six
sniffio
soupsieve
SQLAlchemy
striprtf
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tqdm
typing-inspect
typing_extensions
tzdata
urllib3
wrapt
yarl